move helper files into setup where appropriate
