+++ contracts/marketplace/src/testing/helpers.rs
@@ -1,4 +1,2 @@
-pub mod accounts;
 pub mod funds;
-pub mod msg;
 pub mod nft_functions;
