+++ contracts/marketplace/src/testing/setup.rs
@@ -1,5 +1,7 @@
 pub mod constants;
 pub mod mock_collection_params;
+pub mod msg;
+pub mod setup_accounts;
 pub mod setup_accounts_and_block;
 pub mod setup_contracts;
 pub mod setup_marketplace;
