+++ contracts/marketplace/src/testing/setup/setup_marketplace.rs
@@ -1,9 +1,9 @@
 use crate::error::ContractError;
-use crate::testing::helpers::msg::{MinterCollectionResponse, SetupContractsParams};
 use crate::testing::setup::constants::{
     BID_REMOVAL_REWARD_BPS, LISTING_FEE, MAX_EXPIRY, MAX_FINDERS_FEE_BPS, MIN_EXPIRY,
     TRADING_FEE_BPS,
 };
+use crate::testing::setup::msg::{MinterCollectionResponse, SetupContractsParams};
 use crate::testing::setup::setup_contracts::contract_marketplace;
 use crate::testing::setup::setup_minter::configure_minter;
 use crate::ExpiryRange;
